Refactoring defmodules.
